renaming onboarding -> invitation
