+++ coreservices/CatenaX.NetworkServices.sln
@@ -31,6 +31,12 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "mailing", "mailing", "{F5E5
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "CatenaX.NetworkServices.Mailing.SendMail", "mailing\src\CatenaX.NetworkServices.Mailing.SendMail\CatenaX.NetworkServices.Mailing.SendMail.csproj", "{59FC0798-2A3B-4EAA-B4F2-43B1DAB259DB}"
 EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "CatenaX.NetworkServices.Invitation.Identity", "invitation\src\CatenaX.NetworkServices.Invitation.Identity\CatenaX.NetworkServices.Invitation.Identity.csproj", "{1D6136F2-7946-47DE-95D8-05D021DBA3B4}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "CatenaX.NetworkServices.Invitation.Library", "invitation\src\CatenaX.NetworkServices.Invitation.Library\CatenaX.NetworkServices.Invitation.Library.csproj", "{91955C31-F283-4963-A626-9743BC32AD86}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "CatenaX.NetworkServices.Invitation.Service", "invitation\src\CatenaX.NetworkServices.Invitation.Service\CatenaX.NetworkServices.Invitation.Service.csproj", "{A1088127-BA35-4B52-9F5A-F5B8FE9DAF18}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -61,6 +67,18 @@ Global
 		{59FC0798-2A3B-4EAA-B4F2-43B1DAB259DB}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{59FC0798-2A3B-4EAA-B4F2-43B1DAB259DB}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{59FC0798-2A3B-4EAA-B4F2-43B1DAB259DB}.Release|Any CPU.Build.0 = Release|Any CPU
+		{1D6136F2-7946-47DE-95D8-05D021DBA3B4}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{1D6136F2-7946-47DE-95D8-05D021DBA3B4}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{1D6136F2-7946-47DE-95D8-05D021DBA3B4}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{1D6136F2-7946-47DE-95D8-05D021DBA3B4}.Release|Any CPU.Build.0 = Release|Any CPU
+		{91955C31-F283-4963-A626-9743BC32AD86}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{91955C31-F283-4963-A626-9743BC32AD86}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{91955C31-F283-4963-A626-9743BC32AD86}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{91955C31-F283-4963-A626-9743BC32AD86}.Release|Any CPU.Build.0 = Release|Any CPU
+		{A1088127-BA35-4B52-9F5A-F5B8FE9DAF18}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{A1088127-BA35-4B52-9F5A-F5B8FE9DAF18}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{A1088127-BA35-4B52-9F5A-F5B8FE9DAF18}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{A1088127-BA35-4B52-9F5A-F5B8FE9DAF18}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -78,6 +96,9 @@ Global
 		{97FE948B-D083-4BE2-9E5A-2AA0A0D7B44D} = {F5E59460-211B-11EC-B6E9-75CD881C6D34}
 		{F5E59460-211B-11EC-B6E9-75CD881C6D34} = {C34FEF41-7935-481B-BD92-BB2A99199BE3}
 		{59FC0798-2A3B-4EAA-B4F2-43B1DAB259DB} = {F5E59460-211B-11EC-B6E9-75CD881C6D34}
+		{1D6136F2-7946-47DE-95D8-05D021DBA3B4} = {7A5F84BE-3B45-483A-8EF1-4E1E94F144DC}
+		{91955C31-F283-4963-A626-9743BC32AD86} = {7A5F84BE-3B45-483A-8EF1-4E1E94F144DC}
+		{A1088127-BA35-4B52-9F5A-F5B8FE9DAF18} = {7A5F84BE-3B45-483A-8EF1-4E1E94F144DC}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {23E9EF5A-FB26-4AAC-952E-0695A1245B8E}
