+++ coreservices/invitation/src/CatenaX.NetworkServices.Invitation.Identity/Model/CreateGroup.cs
@@ -1,0 +1,11 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace CatenaX.NetworkServices.Invitation.Identity.Model
+{
+    public class CreateGroup
+    {
+        public string Name { get; set; }
+    }
+}
