+++ coreservices/invitation/src/CatenaX.NetworkServices.Invitation.Identity/Model/CreateRealm.cs
@@ -1,0 +1,11 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace CatenaX.NetworkServices.Invitation.Identity.Model
+{
+    public class CreateRealm
+    {
+        public string Name { get; set; }
+    }
+}
