+++ coreservices/invitation/src/CatenaX.NetworkServices.Invitation.Service/Mail/IMailingService.cs
@@ -1,0 +1,9 @@
+﻿using System.Threading.Tasks;
+
+namespace CatenaX.NetworkServices.Invitation.Service.Mail
+{
+    public interface IMailingService
+    {
+        Task SendMails(string eMail, string password, string companyName);
+    }
+}
