+++ coreservices/invitation/src/CatenaX.NetworkServices.Invitation.Service/appsettings.Development.json
