+++ coreservices/invitation/src/CatenaX.NetworkServices.Invitation.Service/appsettings.json
@@ -1,0 +1,30 @@
+{
+  "Logging": {
+    "LogLevel": {
+      "Default": "Information",
+      "Microsoft": "Warning",
+      "Microsoft.Hosting.Lifetime": "Information"
+    }
+  },
+  "AllowedHosts": "*",
+  "KeyCloakConnectionString": "http://localhost:8080",
+  "KeyCloakUser": "",
+  "KeyCloakUserPassword": "",
+  "MailingService": {
+    "Templates": {
+      "invite": {
+        "Subject": "invitation to Catena-X",
+        "Body": "your company {companyname} is hereby invited to participate in Catena-X"
+      },
+      "password": {
+        "Subject": "password for Invitation to Catena-X",
+        "Body": "please use this password to onboard to Catena-X: {password}"
+      }
+    },
+    "Mail": {
+      "SmtpHost": "",
+      "SmtpUser": "",
+      "SmtpPassword": ""
+    }
+  }
+}
