--- coreservices/invitation/src/CatenaX.NetworkServices.Onboarding.Identity/Model/CreateGroup.cs
@@ -1,11 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
-
-namespace CatenaX.NetworkServices.Onboarding.Identity.Model
-{
-    public class CreateGroup
-    {
-        public string Name { get; set; }
-    }
-}
