--- coreservices/invitation/src/CatenaX.NetworkServices.Onboarding.Library/OnboardingData.cs
@@ -1,16 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
-using System.Text.Json.Serialization;
-
-namespace CatenaX.NetworkServices.Onboarding.Library
-{
-    public class OnboardingData
-    {
-        [JsonPropertyName("oneId")]
-        public string OneId { get; set; }
-
-        [JsonPropertyName("eMail")]
-        public string EMail { get; set; }
-    }
-}
