--- coreservices/invitation/src/CatenaX.NetworkServices.Onboarding.Service/BusinessLogic/IOnboardingBusinessLogic.cs
@@ -1,16 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
-
-using CatenaX.NetworkServices.Onboarding.Library;
-
-namespace CatenaX.NetworkServices.Onboarding.Service.BusinessLogic
-{
-    public interface IOnboardingBusinessLogic
-    {
-        Task ExecuteOnboarding(string identifier);
-
-        Task StartOnboarding(OnboardingData onboardingData);
-    }
-}
