--- coreservices/invitation/src/CatenaX.NetworkServices.Onboarding.Service/CatenaX.NetworkServices.Onboarding.Service.csproj
@@ -1,24 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk.Web">
-
-  <PropertyGroup>
-    <TargetFramework>netcoreapp3.1</TargetFramework>
-    <UserSecretsId>ad1c79a3-a042-4d59-8924-475f8ff4c37a</UserSecretsId>
-    <DockerDefaultTargetOS>Linux</DockerDefaultTargetOS>
-    <DockerfileContext>..\..\..</DockerfileContext>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <PackageReference Include="Dapper" Version="2.0.90" />
-    <PackageReference Include="Microsoft.VisualStudio.Azure.Containers.Tools.Targets" Version="1.11.1" />
-    <PackageReference Include="Npgsql" Version="5.0.10" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\..\..\mailing\src\CatenaX.NetworkServices.Mailing.SendMail\CatenaX.NetworkServices.Mailing.SendMail.csproj" />
-    <ProjectReference Include="..\..\..\mailing\src\CatenaX.NetworkServices.Mailing.Template\CatenaX.NetworkServices.Mailing.Template.csproj" />
-    <ProjectReference Include="..\CatenaX.NetworkServices.Onboarding.Identity\CatenaX.NetworkServices.Onboarding.Identity.csproj" />
-    <ProjectReference Include="..\CatenaX.NetworkServices.Onboarding.Library\CatenaX.NetworkServices.Onboarding.Library.csproj" />
-  </ItemGroup>
-
-
-</Project>
