--- coreservices/invitation/src/CatenaX.NetworkServices.Onboarding.Service/DataAccess/DataContract.cs
@@ -1,14 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
-
-namespace CatenaX.NetworkServices.Onboarding.Service.DataAccess
-{
-    public class DataContract
-    {
-        public int personId { get; set; }
-        public string lastname { get; set; }
-        public string city { get; set; }
-    }
-}
