--- coreservices/invitation/src/CatenaX.NetworkServices.Onboarding.Service/DataAccess/IDataAccess.cs
@@ -1,12 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
-
-namespace CatenaX.NetworkServices.Onboarding.Service.DataAccess
-{
-    public interface IDataAccess
-    {
-        public Task<DataContract> GetData(int id);
-    }
-}
