--- coreservices/invitation/src/CatenaX.NetworkServices.Onboarding.Service/Mail/IMailingService.cs
@@ -1,9 +1,0 @@
-﻿using System.Threading.Tasks;
-
-namespace CatenaX.NetworkServices.Onboarding.Service.Mail
-{
-    public interface IMailingService
-    {
-        Task SendMails(string eMail, string password, string companyName);
-    }
-}
