+++ coreservices/invitation/test/CatenaX.NetworkServices.Onboarding.Service.Tests/OnboardingControllerTest.cs
@@ -2,8 +2,8 @@ using System;
 using System.Net;
 using System.Threading.Tasks;
 
-using CatenaX.NetworkServices.Onboarding.Service.BusinessLogic;
-using CatenaX.NetworkServices.Onboarding.Service.Controllers;
+using CatenaX.NetworkServices.Invitation.Service.BusinessLogic;
+using CatenaX.NetworkServices.Invitation.Service.Controllers;
 
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.Extensions.Logging.Abstractions;
@@ -13,20 +13,20 @@ using NSubstitute.ExceptionExtensions;
 
 using Xunit;
 
-namespace CatenaX.NetworkServices.Onboarding.Service.Tests
+namespace CatenaX.NetworkServices.Invitation.Service.Tests
 {
-    public class OnboardingControllerTest
+    public class InvitationControllerTest
     {
         [Fact]
-        public async Task ExecuteOnboarding_ReturnsInternalServerError_OnException()
+        public async Task ExecuteInvitation_ReturnsInternalServerError_OnException()
         {
             //Setup
-            var onboardingBusinessLogic = Substitute.For<IOnboardingBusinessLogic>(); 
-            onboardingBusinessLogic.ExecuteOnboarding(Arg.Is("testIdent")).Throws<ArgumentException>();
-            var sut = new OnboardingController(NullLogger<OnboardingController>.Instance, onboardingBusinessLogic);
+            var InvitationBusinessLogic = Substitute.For<IInvitationBusinessLogic>(); 
+            InvitationBusinessLogic.ExecuteInvitation(Arg.Is("testIdent")).Throws<ArgumentException>();
+            var sut = new InvitationController(NullLogger<InvitationController>.Instance, InvitationBusinessLogic);
             
             //Execute
-            var result = await sut.ExecuteOnboarding("testIdent");
+            var result = await sut.ExecuteInvitation("testIdent");
             Assert.IsType<StatusCodeResult>(result);
             var statusCodeResult = result as StatusCodeResult;
             Assert.Equal(((int)HttpStatusCode.InternalServerError).ToString(), statusCodeResult.StatusCode.ToString());
